add error handling for invalid ID's to controllers

+++ API/controllers/controllers.go
@@ -18,7 +18,13 @@ func GetAllPresets(c echo.Context) error {
 // GetPresetByID function
 func GetPresetByID(c echo.Context) error {
 	id := c.Param("id")
+	if !bson.IsObjectIdHex(id) {
+		return echo.NewHTTPError(http.StatusBadRequest, "Error: Invalid ObjectId")
+	}
 	res, _ := dao.GetByID(id)
+	if res.ID == "" {
+		return echo.NewHTTPError(http.StatusNotFound, "Error: Unable to find preset")
+	}
 	return c.JSON(http.StatusOK, res)
 }
 
@@ -34,6 +40,9 @@ func CreatePreset(c echo.Context) error {
 // UpdatePreset function
 func UpdatePreset(c echo.Context) error {
 	id := c.Param("id")
+	if !bson.IsObjectIdHex(id) {
+		return echo.NewHTTPError(http.StatusBadRequest, "Error: Invalid ObjectId")
+	}
 	p := new(models.Preset)
 	c.Bind(p)
 	res, _ := dao.UpdatePreset(id, *p)
@@ -43,6 +52,9 @@ func UpdatePreset(c echo.Context) error {
 // DeletePreset function
 func DeletePreset(c echo.Context) error {
 	id := c.Param("id")
+	if !bson.IsObjectIdHex(id) {
+		return echo.NewHTTPError(http.StatusBadRequest, "Error: Invalid ObjectId")
+	}
 	dao.DeletePreset(id)
 	return c.String(http.StatusOK, "Preset deleted!")
 }
